requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.6.1

langchain>=0.3.18
langchain-core>=0.1.27
langchain-community>=0.3.17
langchain-openai>=0.3.8
langchain-deepseek-official
langgraph>=0.3.8
langgraph-prebuilt>=0.1.3

sse-starlette~=2.2.1

ruff

